% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-call-api.R
\name{abs_api_call}
\alias{abs_api_call}
\title{Create ABS.Stat API URL call}
\usage{
abs_api_call(path, args)
}
\arguments{
\item{path}{Character vector specifying one or more ABS collections or catalogue numbers to
download.}

\item{args}{Named list of arguments to supply to call.}
}
\value{
data frame in long format
}
\description{
The function created the ABS.Stat API call URL
}
\seealso{
Other ABS.Stat API call functions: 
\code{\link{abs_api_urls}()},
\code{\link{abs_call_api}()}
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS.Stat API call functions}
\keyword{internal}
