% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-call-api.R
\name{abs_api_urls}
\alias{abs_api_urls}
\title{ABS URL addresses and paths used in ABS.Stat API calls}
\usage{
abs_api_urls()
}
\value{
a list with a base url and a url section for formatting the JSON API calls
}
\description{
This function returns a list of URLs and data paths used to construction ABS.Stat
API call. It is used in other functions in this package and need not be called directly.
}
\seealso{
Other ABS.Stat API call functions: 
\code{\link{abs_api_call}()},
\code{\link{abs_call_api}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS.Stat API call functions}
\keyword{internal}
