% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_cache}
\alias{abs_cache}
\title{Download updated list of datasets and dimensions information from the ABS API}
\usage{
abs_cache(lang = "en", progress = 10)
}
\arguments{
\item{lang}{Language in which to return the results. If \code{lang} is unspecified, English
('en') is the default.}

\item{progress}{Report download progress. Arguments accepts integer, logical or NULL. Set
\code{progress} to \code{NULL} (default) to disable progress
reporting. Otherwise set progress equal to integer value frequency.}
}
\value{
A list of available ABS data series each comprising a list of available data dimensions,
  typically containing:
  \itemize{
    \item \code{MEASURE}: Measurement units (e.g. Persons, $ million, Index, Percentage change, etc.)
    \item \code{REGION}: Australian region name
    \item \code{INDEX}: Data item code and description
    \item \code{TSEST}: Time series estimate type (e.g. Original, Seasonally Adjusted, etc.)
    \item \code{FREQUENCY}: Available data frequency (Monthly, Quarterly, Annual)
    \item \code{TIME}: Available observation period index
    \item \code{OBS_STATUS}: Observation status notes code and description
      (e.g. 'r' - revised, 'q' - not available, 'u' - not applicable)
    \item \code{TIME_FORMAT}: Available time format (e.g. Annual, Quarterly, Monthly, Daily).
  }
}
\description{
TBC
}
\note{
Saving the results of this function and using it as the cache parameter in \code{abs_stats}
  and \code{abs_search} replaces the default cached version \code{abs_cachelist} that comes with
  the package. Note, however, that this function can take a long time to extract metadata for all
  ABS datasets (e.g. approximately 20 minutes for 400 data sets), so use sparingly. For this
  reason, we also recommend specifying a progress update using the \code{progress} argument
  (default: 10).
  Not all data returns have support for languages other than english. If the specific
  return does not support your requested language by default it will return NA. The options for
  \code{lang} on the ABS API are presently:
  \itemize{
    \item en: English
    \item fr: French
  }
}
\examples{
 \dontrun{
   z <- abs_cache(lang='en', progress=5)
 }
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
