% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abs_cachelist}
\alias{abs_cachelist}
\title{Datasets available through the ABS API}
\format{
A data frame containing three columns:
\itemize{
\item \code{id} ABS dataset identifier.
\item \code{agencyID} Source agency identifier (ABS).
\item \code{name} ABS dataset name.
}
}
\usage{
abs_cachelist
}
\description{
This data set provides a list of all datasets, and the associated metadata,
  available through the ABS API.
}
\keyword{datasets}
