% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_cachelist2table}
\alias{abs_cachelist2table}
\title{Converts an abs_cachelist to abs_cachetable}
\usage{
abs_cachelist2table(cache)
}
\arguments{
\item{cache}{An existing cachelist of available ABS datasets created by \code{abs_cachelist}. If
\code{NULL}, uses the stored package cachelist.}
}
\value{
A table containing three columns:
  \itemize{
    \item \code{dataset}: ABS API dataset identifier.
    \item \code{dataset_description}: ABS API dataset description.
    \item \code{measure}: ABS API dataset measure identifier.
    \item \code{measure_description}: ABS API dataset measure description
  }
}
\description{
This function converts an \code{abs_cachelist} to an \code{abs_cachetable} suitable
  for use with \code{\link{abs_search}}.
}
\note{
This is an internal library function and is not exported.
}
\examples{
 \\donotrun{
   abs_ct <- abs_cachelist2table(abs_cachelist)
 }
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
