% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-call-api.R
\name{abs_call_api}
\alias{abs_call_api}
\title{Submit API call to ABS.Stat}
\usage{
abs_call_api(url)
}
\arguments{
\item{url}{Character vector specifying one or more ABS collections or catalogue numbers to
download.}
}
\value{
data frame in long format
}
\description{
This function submits the specified API call to ABS.Stat
}
\seealso{
Other ABS.Stat API call functions: 
\code{\link{abs_api_call}()},
\code{\link{abs_api_urls}()}
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS.Stat API call functions}
\keyword{internal}
