% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abs_cat_cachelist}
\alias{abs_cat_cachelist}
\title{List of ABS catalogue tables}
\format{
A data frame containing five columns:
\itemize{
  \item \code{publication_title} ABS publication title.
  \item \code{catalogue_no} ABS catalogue number.
  \item \code{abs_url} ABS URL.
  \item \code{last_updated} Publication last updated.
  \item \code{type} Publication type -- one of either 'time series', 'panel' or 'summary'.RBA URL 
}
}
\usage{
abs_cat_cachelist
}
\description{
This data set provides a list of the most common ABS catalogue tables.
}
\keyword{datasets}
