% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_cat_download}
\alias{abs_cat_download}
\title{Function to download files from the ABS website and store locally}
\usage{
abs_cat_download(data_urls, exdir = tempdir())
}
\arguments{
\item{data_urls}{Character vector specifying one or more ABS data URLs.}

\item{exdir}{Target directory for downloaded files (defaults to \code{tempdir()}). Directory is
created if it doesn't exist.}
}
\value{
Downloads data from the ABS website and returns a character vector listing the location
  where files are saved.
}
\description{
This function downloads the
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
