% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-select.R
\name{abs_cat_select}
\alias{abs_cat_select}
\title{Return list of ABS statistics series}
\usage{
abs_cat_select(
  pattern,
  level = c("group", "view", "topic"),
  ignore.case = TRUE,
  include_urls = FALSE,
  ...
)
}
\arguments{
\item{pattern}{character string containing a regular expression to be matched in the given
character vector. See \code{\link[base]{grep}} for further details.}

\item{level}{one or more of 'group', 'view' and/or 'topic'.}

\item{...}{other arguments to \code{grep}.}
}
\value{
Returns a data frame listing available ABS catalogue releases.
}
\description{
Return list of ABS statistical series matching selection condition
}
\examples{
  \donttest{
    ## List all available quarterly National Accounts tables
    ana_releases <- abs_cat_releases("5206.0");
    ana_release_urls <- abs_cat_releases("5206.0", include_urls=TRUE);
  
    ## List latest available CPI Time Series Spreadsheet tables only
    cpi_releases <- abs_cat_releases("6401.0");
    cpi_release_urls <- abs_cat_releases("6401.0", include_urls=TRUE);
  }
}
\seealso{
Other ABS catalogue functions: 
\code{\link{abs_cat_download}()},
\code{\link{abs_cat_stats}()},
\code{\link{abs_cat_tables}()},
\code{\link{abs_expressions}()},
\code{\link{abs_read_tss}()},
\code{\link{abs_urls}()}
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS catalogue functions}
