% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-stats.R
\name{abs_cat_stats}
\alias{abs_cat_stats}
\title{Get ABS catalogue series data}
\usage{
abs_cat_stats(
  title,
  cat_no,
  tables = "All",
  releases = "Latest",
  types = "tss",
  na.rm = TRUE
)
}
\arguments{
\item{title}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Character vector specifying one or more ABS publication titles.}

\item{cat_no}{Character vector specifying one or more ABS collections or catalogue numbers to
download.}

\item{tables}{A character vector of regular expressions denoting tables to download. The default
('All') downloads all time series spreadsheet tables for each specified catalogue. Use a list
to specify different table sets for each specified ABS catalogue number.}

\item{releases}{Date or character string object specifying the month and year denoting which
release to download. Default is "Latest", which downloads the latest available data. See
examples for further details.}

\item{types}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
One of either 'tss' -- ABS time series spreadsheet (the default) or 'css' -- ABS
data cube (cross-section spreadsheet).}

\item{na.rm}{logical (default: \code{TRUE}) - remove observations containing missing values.}
}
\value{
data frame in long format
}
\description{
This function downloads ABS catalogue series statistics, by ABS catalogue number.
}
\examples{
  \donttest{
    ## Download quarterly Australian National Accounts, Tables 1 & 2 
    xx <- abs_cat_series(pattern="national.*income.*expenditure.*product");
    ana_q <- abs_cat_stats("5206.0", tables=c("Table 1\\\\W+", "Table 2\\\\W+"));
    ana_q <- abs_cat_stats("5206.0", tables=c("Table 1\\\\W+", "Table 2\\\\W+"));

    ## Download December 2017 Australian National Accounts, Table 1
    ana_q_2017q4 <- abs_cat_stats("5206.0", tables="Table 1\\\\W+", release="Dec 2017");
  }
}
\seealso{
Other ABS catalogue functions: 
\code{\link{abs_cat_download}()},
\code{\link{abs_cat_list}()},
\code{\link{abs_cat_releases}()},
\code{\link{abs_cat_series}()},
\code{\link{abs_cat_tables}()},
\code{\link{abs_expressions}()},
\code{\link{abs_read_tss}()},
\code{\link{abs_urls}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS catalogue functions}
