% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_cat_tables}
\alias{abs_cat_tables}
\title{Return ABS catalogue tables}
\usage{
abs_cat_tables(cat_no, releases = "Latest", include_urls = FALSE)
}
\arguments{
\item{cat_no}{ABS catalogue numbers.}

\item{releases}{Date or character string object specifying the month and year denoting which
release to download. Default is "Latest", which downloads the latest available data. See
examples for further details.}

\item{include_urls}{Return data URLs only, no data (Default: FALSE).}
}
\value{
Returns a data frame listing the data collection tables and links.
}
\description{
Return list of tables from specified ABS catalogue number
}
\examples{
   url <- abs_cat_stats("5206.0", tables=c("Table 1"));
   tables <- abs_cat_tables(url);
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
