% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-dimensions.R
\name{abs_dimensions}
\alias{abs_dimensions}
\title{Return available dimensions of ABS series}
\usage{
abs_dimensions(dataset)
}
\arguments{
\item{dataset}{Character vector of dataset codes. These codes correspond to the
\code{indicatorID} column from the indicator data frame of \code{abs_cache} or
\code{abs_cachelist}, or the result of \code{abs_indicators}.}
}
\value{
a data frame with available dataset dimensions.
}
\description{
This function returns the available dimeninsions for a specified ABS API dataset.
}
\examples{
  \donttest{
    ## CPI - Consumer Price Index
    x <- abs_dimensions("CPI");
    str(x)
    ## LF - Labour Force
    x <- abs_dimensions("LF");
    str(x)
  }
}
\seealso{
Other ABS.Stat functions: 
\code{\link{abs_datasets}()},
\code{\link{abs_metadata}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS.Stat functions}
