% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_download_data}
\alias{abs_download_data}
\title{Function to download files from the ABS website and store locally}
\usage{
abs_download_data(data_urls)
}
\arguments{
\item{data_urls}{Character vector specifying one or more ABS data
URLs.}
}
\value{
Downloads data from the ABS website and returns a character
    vector listing the location where files are saved.
}
\description{
TBC
}
\examples{

}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
