% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-urls.R
\name{abs_expressions}
\alias{abs_expressions}
\title{Common expressions used in scraping the ABS website}
\usage{
abs_expressions()
}
\value{
a list with a set of XPath expressions.
}
\description{
This function returns a common set of expressions used in scraping the ABS website.
}
\seealso{
Other ABS catalogue functions: 
\code{\link{abs_cat_download}()},
\code{\link{abs_cat_list}()},
\code{\link{abs_cat_select}()},
\code{\link{abs_cat_stats}()},
\code{\link{abs_cat_tables}()},
\code{\link{abs_read_tss}()},
\code{\link{abs_urls}()}
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS catalogue functions}
\keyword{internal}
