% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-local-filename.R
\name{abs_local_filename}
\alias{abs_local_filename}
\title{Create local file names for storing downloaded ABS data files}
\usage{
abs_local_filename(url)
}
\arguments{
\item{url}{Character vector specifying one or more ABS data URLs.}
}
\value{
Returns a local file names (character vector) in which downloaded files will be saved.
}
\description{
Function to create local filename from web-based file name.
}
\seealso{
Other ABS catalogue helper functions: 
\code{\link{abs_get_new_site_tables}()},
\code{\link{abs_get_old_site_tables}()},
\code{\link{abs_next_page}()},
\code{\link{abs_read_tss}()},
\code{\link{is_abs_tss}()}
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS catalogue helper functions}
\keyword{internal}
