% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_metadata}
\alias{abs_metadata}
\title{Download updated data series information from the ABS API}
\usage{
abs_metadata(id, lang = "en")
}
\arguments{
\item{id}{ABS dataset ID.}

\item{lang}{Preferred language (default 'en' - English).}

\item{include_notes}{Include ABS annotation information for each series.}
}
\value{
data frame in long format
}
\description{
TBC
}
\examples{
  x <- abs_metadata("CPI");
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
