% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-parse-search.R
\name{abs_parse_search}
\alias{abs_parse_search}
\title{Parse ABS search function results}
\usage{
abs_parse_search(s, resource = c("Statistical analysis and data", "Article"))
}
\arguments{
\item{s}{a session object}

\item{resource}{One or more of \code{'Statistical analysis and data'} (the default) or
\code{'Article'}.}
}
\value{
a data frame containg the ABS Search function results
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function parses the results of the ABS search results and returns as a data
frame.
}
\seealso{
Other ABS search helper functions: 
\code{\link{abs_search_date_filter}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS search helper functions}
\keyword{internal}
