% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-data.R
\name{abs_read_tss}
\alias{abs_read_tss}
\alias{abs_read_tss_}
\title{Read ABS time series data file(s)}
\usage{
abs_read_tss(files, catno, tables = NULL, type = "tss")
}
\arguments{
\item{files}{Names of one or more ABS data file [DEPRECATED]}

\item{catno}{Character vector specifying one or more ABS Catalogue
numbers to download.}

\item{tables}{Regular expression denoting tables to download. If
NULL (Default), downloads all time series spreadsheet tables
for each specified catalogue. Use a list to specify different
table sets for each specified ABS catalogue number.}

\item{type}{One of either 'tss' - time series spreadsheet (DEFAULT
or 'dem' - demographic data}

\item{...}{other arguments to ...}
}
\value{
data frame in long format
}
\description{
This function extracts time series data from ABS data files.
}
\examples{
  x <- file.path(DataDir, "5206001_Key_Aggregates.xls");
ABS.5206001 <- abs_read_tss(File);
ABS.5206001 \%>\% as.data.frame \%>\% head;
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
