% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_search}
\alias{abs_search}
\title{Search dataset information from the ABS API}
\usage{
abs_search(pattern, fields = c("dataset", "dataset_description"),
  extra = FALSE, cache)
}
\arguments{
\item{pattern}{Character string or regular expression to be matched.}

\item{fields}{Character vector of column names through which to search.}

\item{extra}{If FALSE, only the indicator ID and short name are returned, if TRUE, all columns of
the cache parameter's indicator data frame are returned.}

\item{cache}{List of data frames returned from \code{abs_cache}. If omitted, \code{abs_cachelist}
is used.}
}
\value{
A data frame with datasets and data items that match the search pattern.
}
\description{
This function finds datasets that match a search term and returns a data
  frame of matching results.
}
\note{
With acknowledgements to \code{wb_search} function.
}
\examples{
  x <- abs_search(pattern = "consumer price index")
  x <- abs_search(pattern = "census")
  x <- abs_search(pattern = "labour force")
  # with regular expression operators
  # 'unemployment' OR 'employment'
  x <- abs_search(pattern = "unemployment|employment")
  # '^labour force$'
  x <- abs_search(pattern = "^labour force$")
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
