% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-search-date-filter.R
\name{abs_search_date_filter}
\alias{abs_search_date_filter}
\title{Create date filter for ABS search function}
\usage{
abs_search_date_filter(filter)
}
\arguments{
\item{filter}{Date range filter -- one of: 'Today', 'Past
week', 'Past fortnight', 'Past month', 'Past 3 months', 'Past 6 months', 'Past year', or single
calendar year (e.g. '2021' or '2020').}
}
\value{
a string containg formatted ABS search date filter
}
\description{
This function creates a valid date range filter to insert into ABS search URL.
}
\details{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}:
}
\seealso{
Other ABS search helper functions: 
\code{\link{abs_parse_search}()}
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS search helper functions}
\keyword{internal}
