% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_unzip_files}
\alias{abs_unzip_files}
\title{Function to download files from the ABS website and store locally}
\usage{
abs_unzip_files(files)
}
\arguments{
\item{files}{One or more local zip files.}
}
\value{
Downloads data from the ABS website and returns a character
    vector listing the location where files are saved.
}
\description{
TBC
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
