% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-series-abbreviation.R
\name{ana_series_abb}
\alias{ana_series_abb}
\title{Table - ABS ANA Series IDs and series names}
\usage{
ana_series_abb(x)
}
\description{
ABS Series ID  | Series abb    | Series name                                     
:---------------|:--------------|:--------------------------------------
 A2304402X      | gdp_cv_sa     | GDP Chain Volume measures: Seasonally Adjusted   
 A2304340C      | gdp_cv_tr     |           "              : Trend
 A2302459A      | gdp_cv_or     |           "              : Original
 A2304408L      | gdi_cv_sa     | Gross Domestic Income, Chain Volume measures: Seasonally Adjusted
 A2304342J      | gdi_cv_tr     |           "              : Trend
 A2302463T      | gdi_cv_or     |           "              : Original
 A2304412C      | gni_cv_sa     | GNI Chain Volume measures: Seasonally Adjusted
 A2304344L      | gni_cv_tr     |           "              : Trend
 A2302464V      | gni_cv_or     |           "              : Original
 A2304414J      | nndi_cv_sa    | NNDI Chain Volume measures: Seasonally Adjusted
 A2304346T      | nndi_cv_tr    |           "              : Trend
 A2302465W      | nndi_cv_or    |           "              : Original
 A2304404C      | gdppc_cv_sa   | GDP per capita Chain volume measures: Seasonally Adjusted
 A2304336L      | gdppc_cv_tr   |           "              : Trend
 A2302459A      | gdppc_cv_or   |           "              : Original
 A2304113C      | gne_cv_sa     | GNE Chain Volume measures: Seasonally Adjusted
 A2304237F      | gne_cv_tr     |           "              : Trend
 A2302514F      | gne_cv_or     |           "              : Original
 A2304111X      | dfd_cv_sa     | DFD Chain Volume measures: Seasonally Adjusted
 A2304235A      | dfd_cv_tr     |           "              : Trend
 A2302519T      | dfd_cv_or     |           "              : Original
 A2304114F      | exp_cv_sa     | Exports Chain Volume measures: Seasonally Adjusted
 A2304238J      | exp_cv_tr     |           "              : Trend
 A2302520A      | exp_cv_or     |           "              : Original
 A2304115J      | imp_cv_sa     | Imports Chain Volume measures: Seasonally Adjusted
 A2304239K      | imp_cv_sa     |           "              : Trend
 A2302521C      | imp_cv_sa     |           "              : Original
}
\details{
Notes
   GDI = GDP - ToT effects       (GDI - Gross Domestic Income)
   NNDI = ??                     (NNDI - Net National Disposable Income)
   DFD = GNE - Inventory change  (DFD - Domestic Final Demand)
}
