% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_state_codes}
\alias{aus_state_codes}
\title{Table of Australian state and territory codes}
\format{
A data frame containing three columns:
\itemize{
  \item \code{state_code} One-digit state code.
  \item \code{state_abb} State/territory abbreviation.
  \item \code{state_name} State/territory name.
}
}
\usage{
aus_state_codes
}
\description{
A list of Australian state and territory codes (including code 0 -- Australia)
}
\keyword{datasets}
