% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.R
\name{capwords}
\alias{capwords}
\title{Capitalise words in character vectors or string}
\usage{
capwords(s, split = c("\\\\W"), strict = FALSE)
}
\arguments{
\item{s}{Character vector or string}

\item{split}{Specifies base month for first quarter. Can be a scalar: 1,2,3 or character object: Jan,Feb,Mar.}

\item{strict}{The input date format. Default is "\%Y-Q\%q".}
}
\value{
Function returns a character vector object.
}
\description{
Function to capitalise strings.
}
\note{
This function
}
\examples{
## TO DO
}
