% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utilities.R
\name{excel2Date}
\alias{excel2Date}
\title{Convert Excel numeric date to R Date object}
\usage{
excel2Date(x)
}
\arguments{
\item{x}{Excel-based date numeric object}
}
\value{
Date object
}
\description{
Function to convert Excel numeric date to R Date object
}
\keyword{internal}
