% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utilities.R
\name{fin_year}
\alias{fin_year}
\title{Create financial year date object}
\usage{
fin_year(date, ending = "Jun")
}
\arguments{
\item{date}{date object}

\item{ending}{character string abbreviation or number denoting ending month of the financial year}
}
\value{
Date object
}
\description{
Function to create a financial year date object
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
