% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utilities.R
\name{finYear}
\alias{finYear}
\title{Create financial year date object}
\usage{
finYear(date, ending = "Jun")
}
\arguments{
\item{date}{date object}

\item{ending}{character string abbreviation or number denoting ending month of the financial year}
}
\value{
Date object
}
\description{
Function to create a financial year date object
}
\examples{
x <- seq.Date(as.Date("2005-06-01"), length=36, by="month");
finYear(x)
}
