% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-utility-functions.R
\name{is_abs_tss}
\alias{is_abs_tss}
\title{Check if file is a valid ABS time series spreadsheet}
\usage{
is_abs_tss(file)
}
\arguments{
\item{file}{An ABS data filename}
}
\value{
logical
}
\description{
This function checks if the specified returns a list of URLs and data paths used to
construct ABS Catalogue data access calls. It is used in other functions in this package and
need not be called directly.
}
\seealso{
Other ABS catalogue helper functions: 
\code{\link{abs_get_new_site_tables}()},
\code{\link{abs_get_old_site_tables}()},
\code{\link{abs_local_filename}()},
\code{\link{abs_next_page}()},
\code{\link{abs_read_tss}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS catalogue helper functions}
\keyword{internal}
