% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utilities.R
\name{last_day}
\alias{last_day}
\title{Set Date object to the last day of the month}
\usage{
last_day(date)
}
\arguments{
\item{date}{date object}
}
\value{
Date object
}
\description{
Function to change the date of a Date object to the last day of the month
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
