% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utilities.R
\name{lastDay}
\alias{lastDay}
\title{Set Date object to the last day of the month}
\usage{
lastDay(date)
}
\arguments{
\item{date}{date object}
}
\value{
Date object
}
\description{
Function to change the date of a Date object to the last day of the month
}
\examples{
Date <- seq.Date(as.Date("2005-06-01"), length=36, by="month");
last_day(Date)
}
