% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raustats.R
\docType{package}
\name{raustats}
\alias{raustats}
\alias{raustats-package}
\title{raustats: An R package for accessing data and statistics from the ABS and RBA websites}
\description{
The raustats package provides structured access to all data and statistics 
available from the Australian Bureau of Statistics and Reserve Bank of Australia
website, as well as draft access to the ABS.Stat - Beta data catalogue API.
}
\details{
To learn more about the raustats package, start with the vignettes:
\code{browseVignettes(package = "raustats")}
}
