% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr-utilities.R
\name{raustats_check_url_available}
\alias{raustats_check_url_available}
\alias{raustats_check_api_available}
\title{raustats check URL available}
\usage{
raustats_check_api_available(url)
}
\arguments{
\item{url}{The base URL to check.}
}
\value{
\code{TRUE} if the API is available, otherwise \code{stop()} is called.
}
\description{
Function to ensure URL calls fail gracefully with an informative message if the
  resource is not available (and not give a check warning nor error).
}
\details{
Check if the given openSenseMap API endpoint is available
}
\note{
Based on code in \code{opensensmapR}  \url{ttps://github.com/sensebox/opensensmapR/blob/f69cf62b2771d5b6ed605c04b7ddd618f5a272c2/R/api.R}{\code{api.R}}.
}
\keyword{internal}
