% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr-utilities.R
\name{raustats_ua}
\alias{raustats_ua}
\title{raustats package user agent}
\usage{
raustats_ua()
}
\value{
a list with a base url and a url section for formatting the JSON API calls
}
\description{
This function specifies the package user agent, and is used inside
  GET/POST function calls
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
