% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-data.R
\name{rba_data}
\alias{rba_data}
\title{Return data files from a specified url}
\usage{
rba_data(table_code)
}
\arguments{
\item{series}{Character vector specifying one or more ABS collections or catalogue numbers to
download.}

\item{tables}{A character vector of regular expressions denoting tables to download. The default
('All') downloads all time series spreadsheet tables for each specified catalogue. Use a list
to specify different table sets for each specified ABS catalogue number.}

\item{releases}{Date or character string object specifying the month and year denoting which
release to download. Default is "Latest", which downloads the latest available data. See
examples for further details.}

\item{type}{One of either 'tss' - time series spreadsheet (DEFALT or 'dem' - demographic data}

\item{...}{other arguments to}
}
\value{
data frame in long format
}
\description{
TBC
}
\examples{
   x <- get_abs_data("3101.0");
   y <- get_abs_data("5206.0", tables=c("Table 1", "Table 2"));

   x <- get_abs_data("5206.0", tables="Table 1", release="Dec 2017");
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
