% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-functions.R
\name{rba_file_download}
\alias{rba_file_download}
\title{Function to download statistics files from the RBA website and store locally}
\usage{
rba_file_download(url, exdir = tempdir())
}
\arguments{
\item{url}{Character vector specifying one or more RBA data set URLs.}

\item{exdir}{Target directory for downloaded files (defaults to \code{tempdir()}). Directory is
created if it doesn't exist.}
}
\value{
Downloads data from the ABS website and returns a character vector listing the location
  where files are saved.
}
\description{
This function downloads one or more RBA data files at the specified by URLs and
  saves a local copy.
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
