% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-data.R
\name{rba_table_cache}
\alias{rba_table_cache}
\title{Download an updated list of data tables available from the RBA website}
\usage{
rba_table_cache()
}
\arguments{
\item{files}{Names of one or more ABS data file}

\item{type}{One of 'ts' - time series or 'dem' - demographic data}
}
\value{
data frame in long format
}
\description{
...
}
\examples{
  rba_tables <- rba_table_cache();
}
