% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rba_tablecache}
\alias{rba_tablecache}
\title{Cached list of statistical tables provided by the RBA}
\format{A data frame containing three columns:
\itemize{
\item \code{table_code} RBA table code.
\item \code{table_name} RBA table name.
\item \code{table_type} One of either current statistical tables, historical data or discontinued data
\item \code{path} RBA URL 
}}
\usage{
rba_tablecache
}
\description{

}
\keyword{datasets}
