% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-functions.R
\name{rba_urls}
\alias{rba_urls}
\title{RBA base URL and data paths}
\usage{
rba_urls()
}
\value{
list of RBA base URL and data paths
}
\description{
This function returns a list of URLs and data paths used to construct RBA data
  access calls. It is used in other functions in this package and need not be called directly.
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
