% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-abs-discards.R
\name{read_abs}
\alias{read_abs}
\alias{readABS}
\title{Read ABS time series data file(s)}
\usage{
readABS(files, type = "tss")
}
\arguments{
\item{files}{Names of one or more ABS data file}

\item{type}{One of 'ts' - time series or 'dem' - demographic data}
}
\value{
data frame in long format
}
\description{
This function extracts time series data from ABS data files.
}
\examples{
    File <- file.path(DataDir, "5206001_Key_Aggregates.xls");
    Data <- read_abs(File);
    Data \%>\% as.data.frame \%>\% head;
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
