% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-utility-functions.R
\name{split_absstat_query}
\alias{split_absstat_query}
\title{Split long ABS Stat query into several smaller chunks}
\usage{
split_absstat_query(url)
}
\arguments{
\item{url}{An ABS Stat JSON SDMX query string}
}
\value{
string or list
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function checks if a draft ABS.Stat query call exceeds the string length limit
or would return more than one million cells and if so, splits the query into smaller chunks
for calling the ABS.Stat site. This function is called by the \code{\link{abs_stats}} function,
it need not be called directly.
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS.Stat helper functions}
\keyword{internal}
