% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr-utilities.R
\name{valid_url}
\alias{valid_url}
\title{Ensure valid URLs}
\usage{
valid_url(url)
}
\arguments{
\item{url}{Input URL}
}
\value{
valid URL.
}
\description{
Function to ensure URLs are well formed.
}
\keyword{internal}
