% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_api_call}
\alias{abs_api_call}
\title{Download updated indicator information from the ABS API}
\usage{
abs_api_call(path, args)
}
\arguments{
\item{path}{Character vector specifying one or more ABS collections or catalogue numbers to
download.}

\item{args}{Named list of arguments to supply to call.}

\item{releases}{Date or character string object specifying the month and year denoting which
release to download. Default is "Latest", which downloads the latest available data. See
examples for further details.}

\item{type}{One of either 'tss' - time series spreadsheet (the default) or 'css' - cross-section
spreadsheet.}
}
\value{
data frame in long format
}
\description{
TBC
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
