% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_api_urls}
\alias{abs_api_urls}
\title{ABS URL addresses and paths used in ABS.Stat API calls}
\usage{
abs_api_urls()
}
\value{
a list with a base url and a url section for formatting the JSON API calls
}
\description{
This function returns a list of URLs and data paths used to construction ABS.Stat
  API call. It is used in other functions in this package and need not be called directly.
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
