% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_api_urls}
\alias{abs_api_urls}
\title{URL chunks to be used in API calls}
\usage{
abs_api_urls()
}
\value{
a list with a base url and a url section for formatting the JSON API calls
}
\description{
This function is called inside other functions in this package.
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
