% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_call_api}
\alias{abs_call_api}
\title{Download specified URL}
\usage{
abs_call_api(url)
}
\arguments{
\item{url}{Character vector specifying one or more ABS collections or catalogue numbers to
download.}
}
\value{
data frame in long format
}
\description{
Submit API call to ABS.Stat
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
