% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-download.R
\name{abs_cat_download}
\alias{abs_cat_download}
\alias{abs_cat_download.default}
\alias{abs_cat_download.cat_table}
\alias{abs_cat_download.character}
\title{Function to download files from the ABS website and store locally}
\usage{
abs_cat_download(x, exdir = tempdir())

\method{abs_cat_download}{default}(x, exdir = tempdir())

\method{abs_cat_download}{cat_table}(x, exdir = tempdir())

\method{abs_cat_download}{character}(x, exdir = tempdir())
}
\arguments{
\item{x}{Either a character vector specifying one or more a valid ABS data URLs or a data frame
returned by \code{\link{abs_cat_tables}} function.}

\item{exdir}{Target directory for downloaded files (defaults to \code{tempdir()}). Directory is
created if it doesn't exist.}
}
\value{
Downloads data from the ABS website and returns a character vector listing the location
of locally saved files.
}
\description{
Downloads specified ABS catalogue data files from the ABS website, using a valid ABS
data set URL.
}
\seealso{
Other ABS catalogue functions: 
\code{\link{abs_cat_list}()},
\code{\link{abs_cat_releases}()},
\code{\link{abs_cat_series}()},
\code{\link{abs_cat_stats}()},
\code{\link{abs_cat_tables}()},
\code{\link{abs_expressions}()},
\code{\link{abs_read_tss}()},
\code{\link{abs_urls}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS catalogue functions}
