% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-list.R
\name{abs_cat_list}
\alias{abs_cat_list}
\title{Returns a list of all ABS statistical releases}
\usage{
abs_cat_list()
}
\description{
This function traverses the ABS website and get the name and location of the latest
  ABS statistical releases.
}
\details{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}:
}
\examples{
  \donttest{
    abs_cat_list();
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS catalogue functions
#' @keywords internal}
