% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-series.R
\name{abs_cat_series}
\alias{abs_cat_series}
\title{Return list of ABS statistics' series titles}
\usage{
abs_cat_series(
  pattern,
  level = c("group", "view", "title"),
  ignore.case = TRUE,
  include_urls = FALSE,
  ...
)
}
\arguments{
\item{pattern}{character string containing a regular expression to be matched in the given
character vector. See \code{\link[base]{grep}} for further details.}

\item{level}{one or more of 'group', 'view' and/or 'title'.}

\item{ignore.case}{if \code{TRUE} (default) pattern matching is not case sensitive and if
\code{FALSE}, matching is case sensitive.}

\item{include_urls}{if \code{TRUE}, the function includes the ABS website URL for the returned
series in the results table, but not if \code{FALSE} (default)..}

\item{...}{other arguments to \code{\link[base]{grep}}.}
}
\value{
Returns a data frame listing available ABS statistical titles.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Return list of ABS statistical series matching
selection condition
}
\examples{
  \donttest{
    ## List ABS series matching regular expression: "national.*income.*expenditure.*product"
    abs_series <- abs_cat_series(pattern="national.*income.*expenditure.*product",
                                 level="title");

    ## List ABS series matching regular expression: "consumer.*price.*index" and include URLs
    abs_series <- abs_cat_series(pattern="consumer.*price.*index.*product",
                                 include_urls=TRUE);
  }
}
\seealso{
Other ABS catalogue functions: 
\code{\link{abs_cat_download}()},
\code{\link{abs_cat_list}()},
\code{\link{abs_cat_releases}()},
\code{\link{abs_cat_stats}()},
\code{\link{abs_cat_tables}()},
\code{\link{abs_expressions}()},
\code{\link{abs_read_tss}()},
\code{\link{abs_urls}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS catalogue functions}
