% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-tables.R
\name{abs_cat_tables}
\alias{abs_cat_tables}
\title{Return list of tables for specified ABS publication}
\usage{
abs_cat_tables(
  title,
  cat_no,
  releases = "Latest",
  types = c("tss", "css"),
  include_urls = FALSE
)
}
\arguments{
\item{title}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
ABS publication title.}

\item{cat_no}{ABS catalogue numbers.}

\item{releases}{Date or character string object specifying the month and year denoting which
release to download. Default is "Latest", which downloads the latest available data. See
examples for further details.}

\item{types}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
ABS publication types to return. Permissable options include one or more of: 'tss'
-- ABS Time Series Spreadsheets, 'css' -- ABS Data Cubes and 'pub' -- ABS Publications. The
default returns all Time Series Spreadsheets and Data Cubes.}

\item{include_urls}{Include full URLs to returned ABS data files. Default (FALSE) does not
include data file URLs.}
}
\value{
Returns a data frame listing the data collection tables and URLs.
}
\description{
Return list of data tables available for specified ABS data series.
}
\examples{
  \donttest{
    ## List latest available quarterly National Accounts tables
    ana_tables <- abs_cat_tables(title="Australian National Accounts: National Income, Expenditure and Product", releases="Latest");
    ana_tables_url <- abs_cat_tables(cat_no="5206.0", releases="Latest", include_urls=TRUE);

    ## List latest available CPI Time Series Spreadsheet tables only
    cpi_tables <- abs_cat_tables("Consumer Price Index, Australia", releases="Latest", include_urls=TRUE);
    cpi_tables_url <- abs_cat_tables(cat_no="6401.0", releases="Latest");
  
    ## List latest available ASGS Volume 3 Data Cubes
    asgs_vol3_tables <- abs_cat_tables(cat_no="1270.0.55.003", releases="Latest", types="css");
    asgs_vol3_tables_url <- abs_cat_tables(cat_no="1270.0.55.003", releases="Latest",
                                           types="css", include_urls=TRUE);
  
    ## List latest available ASGS ANZSIC publications (PDF) files
    anzsic_2006 <- abs_cat_tables(cat_no="1292.0", releases="Latest", types="pub", include_urls=TRUE);
  }
}
\seealso{
Other ABS catalogue functions: 
\code{\link{abs_cat_download}()},
\code{\link{abs_cat_list}()},
\code{\link{abs_cat_releases}()},
\code{\link{abs_cat_series}()},
\code{\link{abs_cat_stats}()},
\code{\link{abs_expressions}()},
\code{\link{abs_read_tss}()},
\code{\link{abs_urls}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS catalogue functions}
