% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_cat_tables}
\alias{abs_cat_tables}
\title{Return ABS catalogue tables}
\usage{
abs_cat_tables(url)
}
\arguments{
\item{url}{Valid ABS data collection URL.}
}
\value{
Returns a data frame listing the data collection tables and links.
}
\description{
Return list of tables from specified ABS catalogue number
}
\examples{
   url <- abs_cat_data("5206.0", tables=c("Table 1"));
   tables <- abs_cat_tables(url);
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
