% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_datasets}
\alias{abs_datasets}
\title{Download updated data series information from the ABS API}
\usage{
abs_datasets(lang = "en", include_notes = FALSE)
}
\arguments{
\item{lang}{Preferred language (default 'en' - English).}

\item{include_notes}{Include ABS annotation information for each series.}
}
\value{
data frame in long format
}
\description{
TBC
}
\examples{
  datasets <- abs_datasets()
  datasets <- abs_datasets(include_notes=TRUE)
The following code extracts the relevant ABS series information from the returned
XML document by first saving the relevant part of the XML document to an R list and
then explicitly extracting the relevant information from specific nodes by name.
A more general recursive process, impervious to name changes would be preferred,
however, it is more complex than simply revising the following code in response to
potential future server-side changes.
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
