% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-api-functions.R
\name{abs_dimensions}
\alias{abs_dimensions}
\title{Return available dimensions of ABS series}
\usage{
abs_dimensions(dataset, cache)
}
\arguments{
\item{dataset}{Character vector of dataset codes. These codes correspond to the
\code{indicatorID} column from the indicator data frame of \code{abs_cache} or
\code{abs_cachelist}, or the result of \code{abs_indicators}.}

\item{cache}{An existing cachelist of available ABS datasets created by \code{abs_cachelist}. If
\code{NULL}, uses the stored package cachelist.}
}
\value{
a data frame with available dataset dimensions.
}
\description{
This function returns the available dimeninsions for a specified ABS API dataset.
}
\note{
@export
}
\examples{
  # CPI - Consumer Price Index
  x <- abs_dimensions("CPI");
  # LF - Labour Force
  x <- abs_dimensions("LF");
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
