% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-data.R
\name{abs_download_data}
\alias{abs_download_data}
\title{Function to download files from the ABS website and store locally}
\usage{
abs_download_data(data_urls)
}
\arguments{
\item{data_urls}{Character vector specifying one or more ABS data
URLs.}
}
\value{
Downloads data from the ABS website and returns a character
    vector listing the location where files are saved.
}
\description{
TBC
}
\examples{
   x <- abs_cat_data("3101.0");
   y <- abs_cat_data("5206.0", tables=c("Table 1", "Table 2"));

   x <- abs_cat_data("5206.0", tables="Table 1", release="Dec 2017");
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
