% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_filetypes}
\alias{abs_filetypes}
\title{Valid ABS file types}
\usage{
abs_filetypes()
}
\value{
a vector containing a list of valid ABS file types.
}
\description{
This function returns a vector of valid ABS file types for using list of URLs and data paths used to construct ABS Catalogue
  data access calls. It is used in other functions in this package and need not be called
  directly.
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
