% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-get-old-site-tables.R
\name{abs_get_old_site_tables}
\alias{abs_get_old_site_tables}
\title{Navigates ABS AUSSTATS style pages to return list of tables for specified ABS publication}
\usage{
abs_get_old_site_tables(url)
}
\arguments{
\item{url}{ABS statistical release URL.}
}
\value{
Returns a data frame listing the data collection tables and URLs.
}
\description{
Return list of data tables available for specified ABS data series.
}
\seealso{
Other ABS catalogue helper functions: 
\code{\link{abs_get_new_site_tables}()},
\code{\link{abs_local_filename}()},
\code{\link{abs_next_page}()},
\code{\link{abs_read_tss}()},
\code{\link{is_abs_tss}()}
}
\author{
David Mitchell \href{mailto:david.pk.mitchell@gmail.com}{david.pk.mitchell@gmail.com}
}
\concept{ABS catalogue helper functions}
\keyword{internal}
