% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_local_filename}
\alias{abs_local_filename}
\title{Create local file names for storing downloaded ABS data files}
\usage{
abs_local_filename(url)
}
\arguments{
\item{url}{Character vector specifying one or more ABS data URLs.}
}
\value{
Returns a local file names (character vector) in which downloaded files will be saved.
}
\description{
Function to create local filename from web-based file name.
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
