% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-next-page.R
\name{abs_next_page}
\alias{abs_next_page}
\title{Parse ABS search function results}
\usage{
abs_next_page(s)
}
\arguments{
\item{s}{a valid session object (see \link[rvest]{html_session}).}
}
\value{
a url to linking to the next page returned by an ABS search results
}
\description{
This function parses the results of the ABS search results and returns as a data
  frame.
\lifecycle{new}
}
\seealso{
Other ABS catalogue helper functions: 
\code{\link{abs_get_new_site_tables}()},
\code{\link{abs_get_old_site_tables}()},
\code{\link{abs_local_filename}()},
\code{\link{abs_read_tss}()},
\code{\link{is_abs_tss}()}
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\concept{ABS catalogue helper functions}
\keyword{internal}
