% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_read_tss}
\alias{abs_read_tss}
\alias{abs_read_tss_}
\title{Read ABS time series data file(s)}
\usage{
abs_read_tss(files, type = "tss")
}
\arguments{
\item{files}{Names of one or more ABS data files}

\item{type}{One of either 'tss' -- ABS Time Series Spreadsheet (the Default)
or 'css' -- Data Cube.}
}
\value{
data frame in long format
}
\description{
This function extracts time series data from ABS data files.
}
\examples{
  x <- abs_read_tss(file.path("data-raw", "5206001_Key_Aggregates.xls"));
  y <- abs_read_tss(file.path("data-raw", c("5206001_Key_Aggregates.xls","5206002_expenditure_volume_measures.xls")));
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
