% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-cat-functions.R
\name{abs_urls}
\alias{abs_urls}
\title{ABS URL addresses and paths used in accessing ABS Catalogue data calls}
\usage{
abs_urls()
}
\value{
a list with a base url and a url section for formatting ABS Catalogue statistics calls
}
\description{
This function returns a list of URLs and data paths used to construct ABS Catalogue
  data access calls. It is used in other functions in this package and need not be called
  directly.
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
