% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-data.R
\name{rba_data}
\alias{rba_data}
\title{Return data for a specified RBA time series}
\usage{
rba_data(table_code, series_type = "statistical tables",
  update_cache = FALSE)
}
\arguments{
\item{table_code}{Character vector specifying one or more ABS collections or catalogue numbers to download.}

\item{series_type}{RBA series type, one of either 'statistical tables', 'historical data' or 'discontinued data'.}

\item{update_cache}{Boolean argument expressing whether to retrieve an updated list of RBA tables or not (Default).}
}
\value{
data frame in long format
}
\description{
Function to get data from a specified RBA time series.
}
\examples{
   x <- rba_data("A1");
   y <- rba_data("A1", update_cache=TRUE);
   
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
