% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-data.R
\name{rba_data}
\alias{rba_data}
\title{Return data for a specified RBA time series}
\usage{
rba_data(table_code, series_type = "statistical tables",
  update_cache = FALSE)
}
\arguments{
\item{series}{Character vector specifying one or more ABS
collections or catalogue numbers to download.}

\item{tables}{A character vector of regular expressions denoting
tables to download. The default ('All') downloads all time
series spreadsheet tables for each specified catalogue. Use a
list to specify different table sets for each specified ABS
catalogue number.}

\item{releases}{Date or character string object specifying the
month and year denoting which release to download. Default is
"Latest", which downloads the latest available data. See
examples for further details.}

\item{type}{One of either 'tss' - time series spreadsheet (the
default) or 'css' - cross-section spreadsheet}

\item{...}{other arguments to}
}
\value{
data frame in long format
}
\description{
Function to get data from a specified RBA time series.
}
\examples{
   x <- rba_data("A1");
   y <- rba_data("A1", update_cache=TRUE);
   
}
\author{
David Mitchell <david.mitchell@infrastructure.gov.au>
}
