% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-data.R
\name{rba_read_tss}
\alias{rba_read_tss}
\title{Import RBA time series data files}
\usage{
rba_read_tss(files, type = "ts")
}
\arguments{
\item{files}{Names of one or more ABS data file}

\item{type}{One of 'ts' - time series or 'dem' - demographic data}
}
\value{
data frame in long format
}
\description{
Import RBA time series data files
}
\examples{
  File <- file.path(DataDir, "5206001_Key_Aggregates.xls");
  ABS.5206001 <- read_abs(File);
  ABS.5206001 \%>\% as.data.frame \%>\% head;
}
