% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-functions.R
\name{rba_read_tss}
\alias{rba_read_tss}
\title{Read RBA time series spreadsheet}
\usage{
rba_read_tss(files)
}
\arguments{
\item{files}{Names of one or more ABS data file}
}
\value{
data frame in long format
}
\description{
Functions to extract data from a specified RBA time series spreadsheet.
}
\examples{
  file <- file.path("data-raw", "a01whist-summary.xls");
  rba_a1_summ <- rba_read_tss(file);
}
