% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-data.R
\name{rba_search}
\alias{rba_search}
\title{Return list of data tables from RBA website}
\usage{
rba_search(pattern, fields = "table_name", update_cache = FALSE)
}
\arguments{
\item{pattern}{Character string or regular expression to be matched}

\item{fields}{Character vector of column names through which to search}

\item{update_cache}{Update table cache (\code{rba_tablecache}), FALSE by default.}
}
\value{
data frame in long format
}
\description{
Function to return a list of all RBA data tables.
}
\examples{
  x <- rba_search(pattern = "A1", fields="table_code", update_cache=TRUE);

}
