% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-data.R
\name{rba_table_cache}
\alias{rba_table_cache}
\title{Return list of RBA tables}
\usage{
rba_table_cache()
}
\arguments{
\item{files}{Names of one or more ABS data file}

\item{type}{One of 'tss' - time series spreadsheets or 'css' -
cross-section spreadsheets}
}
\value{
data frame in long format
}
\description{
Funtion to return an updated list of data tables
    available from the RBA website.
}
\examples{
  rba_tablecache <- rba_table_cache();
}
