% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-functions.R
\name{rba_table_cache}
\alias{rba_table_cache}
\title{Return list of RBA tables}
\usage{
rba_table_cache()
}
\value{
data frame in long format
}
\description{
Function to return an updated list of data tables available from the RBA website.
}
\examples{
  \donttest{
    rba_cachelist <- rba_table_cache();
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
