% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rba_tablecache}
\alias{rba_tablecache}
\title{Cached list of statistical tables provided by the RBA}
\format{A data frame containing three columns:
\itemize{
\item \code{table_code} RBA table code.
\item \code{table_name} RBA table name.
\item \code{table_type} One of either current statistical tables, historical data or discontinued data
\item \code{path} RBA URL 
}}
\usage{
rba_tablecache
}
\description{
This data is a cached result of the \code{\link{rba_table_cache}} function. By default functions \code{\link{rba_search}} and \code{\link{rba_data}} use this data if the \code{update_cache} parameter is \code{TRUE}.
}
\keyword{datasets}
