% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba-functions.R
\name{rba_urls}
\alias{rba_urls}
\title{RBA base URL and data paths}
\usage{
rba_urls()
}
\value{
list of RBA base URL and data paths
}
\description{
List containing RBA base URL and data paths
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
